+++ src/cdk/activation.ts
@@ -11,6 +11,7 @@ import { AppNode } from './explorer/nodes/appNode'
 import { cdk } from './globals'
 import { ConstructNode } from './explorer/nodes/constructNode'
 import { AslVisualizationCDKManager } from './commands/aslVisualizationCDKManager'
+import { renderGraphCommand } from './commands/renderGraph'
 
 /**
  * Activate AWS CDK related functionality for the extension.
@@ -74,10 +75,7 @@ async function registerCdkCommands(context: vscode.ExtensionContext, explorer: A
     const visualizationManager = new AslVisualizationCDKManager(context)
     context.subscriptions.push(
         vscode.commands.registerCommand('aws.cdk.renderStateMachineGraph', async (node: ConstructNode) => {
-            return await visualizationManager.visualizeStateMachine(
-                context.globalState,
-                node
-            )
+            await renderGraphCommand(node, context.globalState, visualizationManager)
         })
     )
 }
